+++ crates/client/rpc/src/methods/trace/trace_transaction.rs
@@ -14,11 +14,10 @@ use sp_blockchain::HeaderBackend;
 use starknet_core::types::TransactionTraceWithHash;
 use starknet_ff::FieldElement;
 
-use super::utils::{
-    get_previous_block_substrate_hash, map_transaction_to_user_transaction, tx_execution_infos_to_tx_trace,
-};
+use super::utils::{map_transaction_to_user_transaction, tx_execution_infos_to_tx_trace};
 use crate::errors::StarknetRpcApiError;
 use crate::madara_backend_client::get_block_by_block_hash;
+use crate::utils::utils::previous_substrate_block_hash;
 use crate::Starknet;
 
 pub async fn trace_transaction<A, BE, G, C, P, H>(
@@ -56,7 +55,7 @@ where
         Some(transaction_hash_to_trace),
     )?;
 
-    let previous_block_substrate_hash = get_previous_block_substrate_hash(starknet, substrate_block_hash)?;
+    let previous_block_substrate_hash = previous_substrate_block_hash(starknet, substrate_block_hash)?;
 
     let fee_token_address = starknet.client.runtime_api().fee_token_addresses(substrate_block_hash).map_err(|e| {
         log::error!("Failed to retrieve fee token address: {e}");
